requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-filter
drf-spectacular
python-decouple
httpx
pytest
pytest-django
